+++ src/com/gubertmc/MazeGeneratorPlugin.java
@@ -1,9 +1,10 @@
 package com.gubertmc;
 
-import com.gubertmc.maze.commands.Pathfinding3dCommand;
-import com.gubertmc.maze.commands.Pathfinding2dCommand;
-import com.gubertmc.maze.commands.TestCommand;
+import com.gubertmc.plugin.commands.Pathfinding3dCommand;
+import com.gubertmc.plugin.commands.Pathfinding2dCommand;
+import com.gubertmc.plugin.commands.TestCommand;
 import org.bukkit.plugin.java.JavaPlugin;
+
 import java.util.Objects;
 
 public class MazeGeneratorPlugin extends JavaPlugin {
@@ -12,15 +13,18 @@ public class MazeGeneratorPlugin extends JavaPlugin {
     public void onEnable() {
         System.out.println("Initializing MazeGenerator Plugin...");
 
-        // Maze command
+        // Maze 2D command
         Pathfinding2dCommand pathfinding2dCommand = new Pathfinding2dCommand(this);
         Objects.requireNonNull(getCommand("astar2d")).setExecutor(pathfinding2dCommand);
         getServer().getPluginManager().registerEvents(pathfinding2dCommand, this);
 
-        // Maze command
+        // Maze 3D command
         Pathfinding3dCommand genMazeCommand3d = new Pathfinding3dCommand(this);
         Objects.requireNonNull(getCommand("astar3d")).setExecutor(genMazeCommand3d);
         getServer().getPluginManager().registerEvents(genMazeCommand3d, this);
+
+        // Button
+        // getServer().getPluginManager().registerEvents(new ControlPlatform(this, 0), this);
 
         // Test command
         TestCommand testCommand = new TestCommand(this);
