--- src/com/gubertmc/maze/commands/Pathfinding3dCommand.java
@@ -1,361 +1,0 @@
-package com.gubertmc.maze.commands;
-
-import com.gubertmc.MazeGeneratorPlugin;
-import com.gubertmc.maze.astar.algorithm3d.Search3D;
-import com.gubertmc.maze.astar.algorithm3d.Simulation3D;
-import org.bukkit.Location;
-import org.bukkit.Material;
-import org.bukkit.block.Block;
-import org.bukkit.block.BlockFace;
-import org.bukkit.command.Command;
-import org.bukkit.command.CommandExecutor;
-import org.bukkit.command.CommandSender;
-import org.bukkit.entity.ItemFrame;
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.Listener;
-import org.bukkit.event.block.Action;
-import org.bukkit.event.block.BlockBreakEvent;
-import org.bukkit.event.player.PlayerInteractEvent;
-import org.bukkit.inventory.ItemStack;
-import org.bukkit.scheduler.BukkitRunnable;
-
-import java.util.Arrays;
-
-import static org.bukkit.Bukkit.getServer;
-
-public class Pathfinding3dCommand implements CommandExecutor, Listener {
-
-    private final MazeGeneratorPlugin plugin;
-
-    private final String[] difficulties = new String[6];
-
-    private int size = 15;
-    private double wallPercentage = .40;
-
-    private long time = 0;
-
-    private final int[] startCoordinate = new int[3];
-    private final int[] endCoordinate = new int[3];
-
-    private int[][][] maze;
-
-    private boolean validMaze;
-    private boolean platformExists;
-    private boolean aStarEnabled = false;
-
-    private final Material CORE_MATERIAL = Material.LIGHT_BLUE_STAINED_GLASS;
-    private final Material WALL_MATERIAL = Material.EMERALD_BLOCK;
-    private final Material PATH_MATERIAL = Material.GREEN_WOOL;
-    private final Material PATH_SPREAD_MATERIAL = Material.BLUE_STAINED_GLASS;
-    private final Material START_POINT_MATERIAL = Material.BEACON;
-    private final Material END_POINT_MATERIAL = Material.BEACON;
-
-    private Location[][][] locations;
-    private Location startButtonLocation;
-
-    private BlockBreakEvent bbe;
-
-    public Pathfinding3dCommand(MazeGeneratorPlugin plugin) {
-        this.plugin = plugin;
-        difficulties[0] = "EASY";
-        difficulties[1] = "SIMPLE";
-        difficulties[2] = "MEDIUM";
-        difficulties[3] = "MODERATE";
-        difficulties[4] = "HARD";
-        difficulties[5] = "DIFFICULT";
-    }
-
-    @Override
-    public boolean onCommand(CommandSender commandSender, Command command, String alias, String[] args) {
-        String difficulty = "MEDIUM";
-        if (args.length > 0 && args.length < 3) {
-            try {
-                size = Integer.parseInt(args[0]);
-                if (args.length == 2) {
-                    difficulty = args[1].toUpperCase();
-                    if (Arrays.asList(difficulties).contains(difficulty)) {
-                        getServer().broadcastMessage("Maze command initiated.");
-                        getServer().broadcastMessage("Break a block to spawn a " + args[0] + "x" + args[0] + " (" + difficulty.toUpperCase() + ") maze.");
-                    } else {
-                        getServer().broadcastMessage("Invalid syntax. Proper syntax is /genmaze <width> <difficulty>");
-                        getServer().broadcastMessage("<width> is required and must be an integer.");
-                        getServer().broadcastMessage("<difficulty> is optional and must be \"easy\", \"medium\", or \"hard\".");
-                    }
-                } else {
-                    getServer().broadcastMessage("Maze command initiated.");
-                    getServer().broadcastMessage("Break a block to spawn a " + args[0] + "x" + args[0] + " (" + difficulty.toUpperCase() + ") maze.");
-                }
-            } catch (NumberFormatException e) {
-                getServer().broadcastMessage("Invalid syntax. Proper syntax is /genmaze <width> <difficulty>");
-                getServer().broadcastMessage("<width> is required and must be an integer.");
-                getServer().broadcastMessage("<difficulty> is optional and must be \"easy\", \"medium\", or \"hard\".");
-            }
-        } else {
-            getServer().broadcastMessage("Invalid syntax. Proper syntax is /genmaze <width> <difficulty>");
-            getServer().broadcastMessage("<width> is required and must be an integer.");
-            getServer().broadcastMessage("<difficulty> is optional and must be \"easy\", \"medium\", or \"hard\".");
-        }
-
-        switch (difficulty) {
-            case "EASY", "SIMPLE" -> wallPercentage = .02;
-            case "MEDIUM", "MODERATE" -> wallPercentage = .06;
-            case "HARD", "DIFFICULT" -> wallPercentage = .10;
-        }
-
-        locations = new Location[size][size][size];
-        aStarEnabled = true;
-        time = 0;
-        return true;
-    }
-
-    @EventHandler
-    public void onBlockBreak(BlockBreakEvent e) {
-        if (aStarEnabled) {
-            getServer().broadcastMessage("Spawned " + size + "x" + size + "x" + size + " maze...");
-            e.isCancelled();
-            aStarEnabled = false;
-            validMaze = false;
-            this.bbe = e;
-            platformExists = true;
-            spawnPlatform(e);
-            generateMazeCore(e);
-        }
-    }
-
-    @EventHandler
-    public void onStartButtonClicked(PlayerInteractEvent e) {
-        if (e.getAction() == Action.RIGHT_CLICK_BLOCK) {
-            Block clicked = e.getClickedBlock();
-            assert clicked != null;
-            if (clicked.getType() == Material.WARPED_BUTTON && startButtonLocation.getBlock().getRelative(BlockFace.WEST).getType() == clicked.getLocation().getBlock().getType()) {
-                time = 0L;
-                if (platformExists) {
-                    getServer().broadcastMessage("Generating new maze...");
-                    generateMazeCore(bbe);
-                }
-                int i = 0;
-                validMaze = false;
-                while (!validMaze) {
-                    Pathfinding3dCommand.this.maze = generateSimulationMaze();
-                    Simulation3D searchSimulation = new Simulation3D(maze, startCoordinate, endCoordinate);
-                    validMaze = searchSimulation.start();
-
-                    if (validMaze) {
-                        generateNewMaze(bbe);
-
-                        time += 10L;
-                        new BukkitRunnable() {
-                            @Override
-                            public void run() {
-                                Search3D search = new Search3D(plugin, locations, startCoordinate, endCoordinate, size, WALL_MATERIAL, PATH_MATERIAL, PATH_SPREAD_MATERIAL, CORE_MATERIAL);
-                                validMaze = search.start();
-                                getServer().broadcastMessage("Maze generated.");
-                                search.showAnimation(time);
-                                cancel();
-                            }
-                        }.runTaskTimer(this.plugin, time, 20L);
-
-                        time = 0;
-                    } else {
-                        getServer().broadcastMessage("Invalid maze - retrying..." + i);
-                        i++;
-                    }
-                }
-            }
-        }
-    }
-
-    public int[][][] generateSimulationMaze() {
-
-        int[][][] maze = new int[size][size][size];
-
-        for (int i = 0; i < size; i++) {
-            for (int j = 0; j < size; j++) {
-                for (int k = 0; k < size; k++) {
-                    maze[i][j][k] = 0;
-                }
-            }
-        }
-
-        int randomStartX = (int) (Math.random() * size);
-        int randomStartY = (int) (Math.random() * size);
-        int randomStartZ = (int) (Math.random() * size);
-
-        int randomEndX = (int) (Math.random() * size);
-        int randomEndY = (int) (Math.random() * size);
-        int randomEndZ = (int) (Math.random() * size);
-
-        boolean badPositions = true;
-        while (badPositions) {
-            randomStartX = (int) (Math.random() * size);
-            randomStartY = (int) (Math.random() * size);
-            randomStartZ = (int) (Math.random() * size);
-            randomEndX = (int) (Math.random() * size);
-            randomEndY = (int) (Math.random() * size);
-            randomEndZ = (int) (Math.random() * size);
-
-            int distance = (int) Math.sqrt(Math.pow(randomEndX - randomStartX, 2) + Math.pow(randomEndY - randomStartY, 2) + Math.pow(randomEndZ - randomStartZ, 2));
-            if (size / 1.3 < distance) {
-                badPositions = false;
-            }
-        }
-
-        maze[randomStartX][randomStartY][randomStartZ] = 4;
-        maze[randomEndX][randomEndY][randomEndZ] = 5;
-
-        startCoordinate[0] = randomStartY;
-        startCoordinate[1] = randomStartX;
-        startCoordinate[2] = randomStartZ;
-
-        endCoordinate[0] = randomEndY;
-        endCoordinate[1] = randomEndX;
-        endCoordinate[2] = randomEndZ;
-
-        int k = 0;
-        int randomX = (int) (Math.random() * size);
-        int randomY = (int) (Math.random() * size);
-        int randomZ = (int) (Math.random() * size);
-        for (int i = 0; i < (size * size * size) * wallPercentage; i++) {
-            if (k % 2 == 0) {
-                randomX = (int) (Math.random() * size);
-                randomY = (int) (Math.random() * size);
-                randomZ = (int) (Math.random() * size);
-            } else {
-                int random = (int) (Math.random() * 3);
-                if (random == 0) {
-                    if (randomX < size - 1 && randomX + 1 != randomStartX && randomX + 1 != randomStartY) {
-                        randomX++;
-                    }
-                } else if (random == 1) {
-                    if (randomY < size - 1 && randomY + 1 != randomStartX && randomY + 1 != randomStartY) {
-                        randomY++;
-                    }
-                } else {
-                    if (randomZ < size - 1 && randomZ + 1 != randomStartZ && randomZ + 1 != randomStartZ) {
-                        randomZ++;
-                    }
-                }
-            }
-            while ((randomX == randomStartX && randomY == randomStartY && randomZ == randomStartZ) || (randomX == randomEndX && randomY == randomEndY && randomZ == randomEndZ)) {
-                randomX = (int) (Math.random() * size);
-                randomY = (int) (Math.random() * size);
-                randomZ = (int) (Math.random() * size);
-            }
-            maze[randomX][randomY][randomZ] = 1;
-            k++;
-        }
-
-        return maze;
-    }
-
-    public void spawnPlatform(BlockBreakEvent e) {
-        for (int i = 0; i < 5; i++) {
-            for (int j = 0; j < 3; j++) {
-                Location floor = new Location(e.getBlock().getWorld(), e.getBlock().getX() + i, e.getBlock().getY(), e.getBlock().getZ() - 4 + j);
-                floor.getBlock().setType(CORE_MATERIAL);
-            }
-            for (int k = 0; k < 2; k++) {
-                Location location = new Location(e.getBlock().getWorld(), e.getBlock().getX() + i, e.getBlock().getY() + 1 + k, e.getBlock().getZ() - 2);
-                location.getBlock().setType(WALL_MATERIAL);
-                if (k == 1 && i != 2) {
-                    ItemFrame frame = e.getBlock().getWorld().spawn(location.add(0, 0, -1), ItemFrame.class);
-                    frame.setItem(new ItemStack(Material.DIAMOND_AXE));
-                }
-                if (k == 1 && i == 3) {
-                    Location loc = new Location(e.getBlock().getWorld(), e.getBlock().getX() + i, e.getBlock().getY() + 1 + k, e.getBlock().getZ() - 3);
-                    loc.getBlock().getRelative(BlockFace.WEST).setType(Material.WARPED_BUTTON);
-                    startButtonLocation = loc.getBlock().getLocation();
-                }
-            }
-        }
-    }
-
-    public void generateNewMaze(BlockBreakEvent e) {
-        generateArenaStartAndEndPoints(e);
-        generateMazeWalls(e);
-    }
-
-    public void generateMazeCore(BlockBreakEvent e) {
-        time += 10L;
-        int count = 0;
-        for (int i = 0; i < size; i++) {
-            for (int j = 0; j < size; j++) {
-                for (int k = 0; k < size; k++) {
-                    Location floor = new Location(e.getPlayer().getWorld(), e.getBlock().getX() + i, e.getBlock().getY() + k, e.getBlock().getZ() + j);
-                    if (floor.getBlock().getType() == CORE_MATERIAL) {
-                        locations[i][j][k] = floor;
-                    } else {
-                        runnableDelayed(floor, time, CORE_MATERIAL, i, j, k);
-                    }
-                    if (count % 90 == 0) { // was 50
-                        time += 1L;
-                    }
-                    count++;
-                }
-            }
-            if (i % 12 == 0) { // was 25
-                time += 15L;
-            }
-        }
-        time += 10L;
-    }
-
-    public void generateArenaStartAndEndPoints(BlockBreakEvent e) {
-        time += 10L;
-        int randomStartX = startCoordinate[1];
-        int randomStartY = startCoordinate[0];
-        int randomStartZ = startCoordinate[2];
-
-        int randomEndX = endCoordinate[1];
-        int randomEndY = endCoordinate[0];
-        int randomEndZ = endCoordinate[2];
-
-        Location startPoint = new Location(e.getPlayer().getWorld(), e.getBlock().getX() + randomStartX, e.getBlock().getY() + randomStartZ, e.getBlock().getZ() + randomStartY);
-        runnableDelayed(startPoint, time, START_POINT_MATERIAL, -1, -1, -1);
-        startCoordinate[0] = randomStartY;
-        startCoordinate[1] = randomStartX;
-        startCoordinate[2] = randomStartZ;
-
-        Location endPoint = new Location(e.getPlayer().getWorld(), e.getBlock().getX() + randomEndX, e.getBlock().getY() + randomEndZ, e.getBlock().getZ() + randomEndY);
-        runnableDelayed(endPoint, time, END_POINT_MATERIAL, -1, -1, -1);
-        endCoordinate[0] = randomEndY;
-        endCoordinate[1] = randomEndX;
-        endCoordinate[2] = randomEndZ;
-
-        time += 10L;
-    }
-
-    public void generateMazeWalls(BlockBreakEvent e) {
-        time += 10L;
-        for (int i = 0; i < size; i++) {
-            for (int j = 0; j < size; j++) {
-                for (int k = 0; k < size; k++) {
-                    if (maze[i][j][k] == 1) {
-                        Location mazeWall = new Location(e.getPlayer().getWorld(), e.getBlock().getX() + i, e.getBlock().getY() + k, e.getBlock().getZ() + j);
-                        runnableDelayed(mazeWall, time, WALL_MATERIAL, i, j, k);
-                    }
-                }
-            }
-            if (i % (int)(size * .15) == 0) {
-                time += 2L;
-            }
-        }
-        time += 10L;
-    }
-
-    public void runnableDelayed(Location loc, long time, Material material, int i, int j, int k) {
-        new BukkitRunnable() {
-            @Override
-            public void run() {
-                if (i < 0 || j < 0 || i > size || j > size || k < 0 || k > size) {
-                    loc.getBlock().setType(material);
-                    cancel();
-                } else {
-                    loc.getBlock().setType(material);
-                    locations[i][j][k] = loc;
-                    cancel();
-                }
-            }
-        }.runTaskTimer(this.plugin, time, 20L);
-    }
-}
