+++ src/com/gubertmc/plugin/algorithms/Search.java
@@ -1,26 +1,58 @@
-package com.gubertmc.maze.astar;
+package com.gubertmc.plugin.algorithms;
 
-import java.util.ArrayList;
-import java.util.PriorityQueue;
+import com.gubertmc.MazeGeneratorPlugin;
+import org.bukkit.Location;
+import org.bukkit.Material;
+import org.bukkit.event.Listener;
+import org.bukkit.scheduler.BukkitRunnable;
 
-public abstract class Simulation {
+import java.util.*;
+
+public class Search implements Listener {
 
     public int SIZE;
+    public Material WALL_MATERIAL;
+    public Material PATH_MATERIAL;
+    public Material PATH_SPREAD_MATERIAL;
+    public Material PATH_GROUND_MATERIAL;
+    public Material START_POINT_GLASS;
+    public Material END_POINT_GLASS;
     public Node[][][] grid;
     public PriorityQueue<Node> open_list = new PriorityQueue<>(10, new NodeComparator());
     public ArrayList<Node> closed_list = new ArrayList<>();
-    public int[][][] tile_grid;
+    public Location[][][] tile_grid;
+    public int[][][] tile_grid_int;
     public Node start_node;
     public Node current_node;
-    public final Node end_node;
+    public Node end_node;
     public boolean is3d;
+    public MazeGeneratorPlugin plugin;
+    public ArrayList<Location> thePath = new ArrayList<>();
+    public ArrayList<Location> exploredPlaces = new ArrayList<>();
 
-    public Simulation(int[][][] maze, int[] startCoordinate, int[] endCoordinate, boolean is3d) {
-        int size = maze[0].length;
-        SIZE = size;
+    public Search(MazeGeneratorPlugin plugin, Location[][][] tiles, int[] startCoordinate, int[] endCoordinate, int size, Material wallMaterial, Material pathMaterial, Material pathSpreadMaterial, Material groundMaterial, Material startGlassMaterial, Material endGlassMaterial, boolean is3d) {
         grid = new Node[size][size][size];
-        tile_grid = maze;
+
+        int[][][] tempArray = new int[size][size][size];
+        for (int i = 0; i < size; i++) {
+            for (int j = 0; j < size; j++) {
+                for (int k = 0; k < size; k++) {
+                    tempArray[i][j][k] = 0;
+                }
+            }
+        }
+        tile_grid_int = tempArray;
+
+        this.plugin = plugin;
         this.is3d = is3d;
+        SIZE = size;
+        WALL_MATERIAL = wallMaterial;
+        PATH_MATERIAL = pathMaterial;
+        PATH_SPREAD_MATERIAL = pathSpreadMaterial;
+        PATH_GROUND_MATERIAL = groundMaterial;
+        START_POINT_GLASS = startGlassMaterial;
+        END_POINT_GLASS = endGlassMaterial;
+        tile_grid = tiles;
 
         if (!is3d) {
             current_node = new Node(startCoordinate[1], startCoordinate[0], -1, 0);
@@ -30,11 +62,11 @@ public abstract class Simulation {
 
             for (int i = 0; i < SIZE; i++) {
                 for (int j = 0; j < SIZE; j++) {
-                    if (tile_grid[i][j][0] == 0) {
+                    if (tiles[i][j][0].getBlock().getType() == Material.AIR) {
                         Node node = new Node(i, j, -1, 0);
                         grid[i][j][0] = node;
                     }
-                    if (tile_grid[i][j][0] == 1) {
+                    if (tiles[i][j][0].getBlock().getType() == WALL_MATERIAL) {
                         Node node = new Node(i, j, -1, 1);
                         grid[i][j][0] = node;
                     }
@@ -49,11 +81,11 @@ public abstract class Simulation {
             for (int i = 0; i < SIZE; i++) {
                 for (int j = 0; j < SIZE; j++) {
                     for (int k = 0; k < SIZE; k++) {
-                        if (tile_grid[i][j][k] == 0) {
+                        if (tiles[i][j][k].getBlock().getType() == PATH_GROUND_MATERIAL) {
                             Node node = new Node(i, j, k, 0);
                             grid[i][j][k] = node;
                         }
-                        if (tile_grid[i][j][k] == 1) {
+                        if (tiles[i][j][k].getBlock().getType() == WALL_MATERIAL) {
                             Node node = new Node(i, j, k, 1);
                             grid[i][j][k] = node;
                         }
@@ -61,7 +93,6 @@ public abstract class Simulation {
                 }
             }
         }
-
         int g = calculateG(current_node);
         current_node.setG(g);
         int h = calculateH(current_node);
@@ -73,9 +104,10 @@ public abstract class Simulation {
 
     public boolean start() {
         boolean pathFound = true;
-        while (!open_list.isEmpty() && !current_node.equals(end_node)) { // open list isn't empty or goal node isn't reached
+        while (!open_list.isEmpty() && !current_node.equals(end_node)) {
             current_node = open_list.peek();
             open_list.remove(open_list.peek());
+
             if (current_node.equals(end_node)) {
                 closed_list.add(current_node);
                 ArrayList<Node> path = generatePath();
@@ -87,26 +119,25 @@ public abstract class Simulation {
                     if (!is3d) {
                         zNum = 0;
                     }
+                    if (tile_grid_int[row][col][zNum] == 1) {
+                        int x = tile_grid[row][col][zNum].getBlockX();
+                        int y;
+                        if (is3d) {
+                            y = tile_grid[row][col][zNum].getBlockY();
+                        } else {
+                            y = tile_grid[row][col][0].getBlockY() - 1;
+                        }
+                        int z = tile_grid[row][col][zNum].getBlockZ();
 
-                    if (tile_grid[row][col][zNum] == 2) {
-                        tile_grid[row][col][zNum] = 3;
+                        Location floor = new Location(tile_grid[row][col][zNum].getWorld(), x, y, z);
+                        thePath.add(floor);
                     }
                 }
                 break;
             } else {
                 try {
                     calculateNeighborValues();
-                } catch (NullPointerException np) {
-                    System.out.println(np.getMessage());
-                }
-
-                if (!is3d) {
-                    tile_grid[start_node.getRow()][start_node.getCol()][0] = 4;
-                    tile_grid[end_node.getRow()][end_node.getCol()][0] = 5;
-                } else {
-                    tile_grid[start_node.getRow()][start_node.getCol()][start_node.getZ()] = 4;
-                    tile_grid[end_node.getRow()][end_node.getCol()][end_node.getZ()] = 5;
-                }
+                } catch (NullPointerException np) {}
 
                 try {
                     assert open_list.peek() != null;
@@ -121,6 +152,42 @@ public abstract class Simulation {
             pathFound = false;
         }
         return pathFound;
+    }
+
+    public void showAnimation(long time) {
+        time += 50L;
+        int count = 1;
+        for (Location loc : exploredPlaces) {
+            runnableDelayed(loc, time, PATH_SPREAD_MATERIAL);
+            count++;
+            if (count % (int) (SIZE * 0.25) == 0) {
+                time += 1L;
+            }
+        }
+
+        time += 10L;
+        for (Location loc : thePath) {
+            if (thePath.get(thePath.size() - 1) == loc) {
+                // do something cool
+            } else {
+                runnableDelayed(loc, time, PATH_MATERIAL);
+                time += 1L;
+            }
+        }
+    }
+
+    public void runnableDelayed(Location loc, long time, Material material) {
+        new BukkitRunnable() {
+            @Override
+            public void run() {
+                if (loc.getBlock().getType() != END_POINT_GLASS && !is3d) {
+                    loc.getBlock().setType(material);
+                } else if (loc.getBlock().getType() != Material.BEACON && is3d) {
+                    loc.getBlock().setType(material);
+                }
+                cancel();
+            }
+        }.runTaskTimer(this.plugin, time, 20L);
     }
 
     public int calculateG(Node node) {
@@ -204,6 +271,17 @@ public abstract class Simulation {
         return x + y + z;
     }
 
+    public ArrayList<Node> generatePath() {
+        ArrayList<Node> path = new ArrayList<>();
+        Node temp = current_node;
+        path.add(temp);
+        while (temp.getParent() != null) {
+            temp = temp.getParent();
+            path.add(temp);
+        }
+        return path;
+    }
+
     public void calculateNeighborValues() {
         int row = current_node.getRow();
         int col = current_node.getCol();
@@ -222,7 +300,17 @@ public abstract class Simulation {
             grid[row - 1][col][zNum].setH(h);
             grid[row - 1][col][zNum].setF();
             open_list.add(grid[row - 1][col][zNum]);
-            tile_grid[row - 1][col][zNum] = 2;
+            tile_grid_int[row - 1][col][zNum] = 1;
+
+            Location loc = tile_grid[row - 1][col][zNum];
+            if (!exploredPlaces.contains(loc)) {
+                if (is3d) {
+                    loc = new Location(loc.getWorld(), loc.getBlock().getX(), loc.getBlock().getY(), loc.getBlock().getZ());
+                } else {
+                    loc = new Location(loc.getWorld(), loc.getBlock().getX(), loc.getBlock().getY() - 1, loc.getBlock().getZ());
+                }
+                exploredPlaces.add(loc);
+            }
         }
 
         // left node
@@ -234,7 +322,17 @@ public abstract class Simulation {
             grid[row][col + 1][zNum].setH(h);
             grid[row][col + 1][zNum].setF();
             open_list.add(grid[row][col + 1][zNum]);
-            tile_grid[row][col + 1][zNum] = 2;
+            tile_grid_int[row][col + 1][zNum] = 1;
+
+            Location loc = tile_grid[row][col + 1][zNum];
+            if (!exploredPlaces.contains(loc)) {
+                if (is3d) {
+                    loc = new Location(loc.getWorld(), loc.getBlock().getX(), loc.getBlock().getY(), loc.getBlock().getZ());
+                } else {
+                    loc = new Location(loc.getWorld(), loc.getBlock().getX(), loc.getBlock().getY() - 1, loc.getBlock().getZ());
+                }
+                exploredPlaces.add(loc);
+            }
         }
 
         // behind node
@@ -246,7 +344,17 @@ public abstract class Simulation {
             grid[row + 1][col][zNum].setH(h);
             grid[row + 1][col][zNum].setF();
             open_list.add(grid[row + 1][col][zNum]);
-            tile_grid[row + 1][col][zNum] = 2;
+            tile_grid_int[row + 1][col][zNum] = 1;
+
+            Location loc = tile_grid[row + 1][col][zNum];
+            if (!exploredPlaces.contains(loc)) {
+                if (is3d) {
+                    loc = new Location(loc.getWorld(), loc.getBlock().getX(), loc.getBlock().getY(), loc.getBlock().getZ());
+                } else {
+                    loc = new Location(loc.getWorld(), loc.getBlock().getX(), loc.getBlock().getY() - 1, loc.getBlock().getZ());
+                }
+                exploredPlaces.add(loc);
+            }
         }
 
         // right node
@@ -258,7 +366,17 @@ public abstract class Simulation {
             grid[row][col - 1][zNum].setH(h);
             grid[row][col - 1][zNum].setF();
             open_list.add(grid[row][col - 1][zNum]);
-            tile_grid[row][col - 1][zNum] = 2;
+            tile_grid_int[row][col - 1][zNum] = 1;
+
+            Location loc = tile_grid[row][col - 1][zNum];
+            if (!exploredPlaces.contains(loc)) {
+                if (is3d) {
+                    loc = new Location(loc.getWorld(), loc.getBlock().getX(), loc.getBlock().getY(), loc.getBlock().getZ());
+                } else {
+                    loc = new Location(loc.getWorld(), loc.getBlock().getX(), loc.getBlock().getY() - 1, loc.getBlock().getZ());
+                }
+                exploredPlaces.add(loc);
+            }
         }
 
         if (is3d) {
@@ -271,7 +389,13 @@ public abstract class Simulation {
                 grid[row][col][zNum - 1].setH(h);
                 grid[row][col][zNum - 1].setF();
                 open_list.add(grid[row][col][zNum - 1]);
-                tile_grid[row][col][zNum - 1] = 2;
+                tile_grid_int[row][col][zNum - 1] = 1;
+
+                Location loc = tile_grid[row][col][zNum - 1];
+                if (!exploredPlaces.contains(loc)) {
+                    loc = new Location(loc.getWorld(), loc.getBlock().getX(), loc.getBlock().getY(), loc.getBlock().getZ());
+                    exploredPlaces.add(loc);
+                }
             }
 
             // top node
@@ -283,19 +407,14 @@ public abstract class Simulation {
                 grid[row][col][zNum + 1].setH(h);
                 grid[row][col][zNum + 1].setF();
                 open_list.add(grid[row][col][zNum + 1]);
-                tile_grid[row][col][zNum + 1] = 2;
+                tile_grid_int[row][col][zNum + 1] = 1;
+
+                Location loc = tile_grid[row][col][zNum + 1];
+                if (!exploredPlaces.contains(loc)) {
+                    loc = new Location(loc.getWorld(), loc.getBlock().getX(), loc.getBlock().getY(), loc.getBlock().getZ());
+                    exploredPlaces.add(loc);
+                }
             }
         }
-    }
-
-    public ArrayList<Node> generatePath() {
-        ArrayList<Node> path = new ArrayList<>();
-        Node temp = current_node;
-        path.add(temp);
-        while (temp.getParent() != null) {
-            temp = temp.getParent();
-            path.add(temp);
-        }
-        return path;
     }
 }
