+++ src/com/gubertmc/plugin/commands/TestCommand.java
@@ -1,4 +1,4 @@
-package com.gubertmc.maze.commands;
+package com.gubertmc.plugin.commands;
 
 import com.gubertmc.MazeGeneratorPlugin;
 import org.bukkit.Location;
