+++ src/plugin.yml
@@ -4,10 +4,10 @@ api-version: 1.13
 version: 1
 commands:
   astar2d:
-    description: creates a maze
+    description: creates a plugin
     usage: /astar2d
   astar3d:
-    description: creates a 3d maze
+    description: creates a 3d plugin
     usage: /astar3d
   test:
     description: test commands
